thêm swagger cho source

+++ Server/controller/SourceController.js
@@ -1,4 +1,5 @@
 const Source = require('../models/Source');
+const Article = require('../models/Article')
 
 class SourceRouter   {
     async getSources(req, res, next) {
@@ -10,7 +11,7 @@ class SourceRouter   {
             });
         }
         catch (err) {
-            res.json({
+            res.status(500).json({
                 success: false,
                 error : err,
             });
@@ -27,7 +28,7 @@ class SourceRouter   {
             });
         }
         catch (err) {
-            res.status(400).json({
+            res.status(500).json({
                 success: false,
                 error : err,
             });
@@ -40,7 +41,7 @@ class SourceRouter   {
 
         try {
             if(await Source.findOne({ url : source.url })) {
-                res.json({
+                res.status(400).json({
                   success : false,
                   message : "Tồn tại trang web khác có cùng url"
                 });
@@ -53,11 +54,10 @@ class SourceRouter   {
             })
         }
         catch (err) {
-            res.status(400).json({
+            res.status(500).json({
                 success: false,
                 error : err,
             })
-            console.log(err);
         }
 
     }
@@ -68,7 +68,7 @@ class SourceRouter   {
         try {
             const check = await Source.findOne({ url : updatedFields.url })
             if(check && (check._id.toString() !== sourceId)){
-                res.json({
+                res.status(400).json({
                     success : false,
                     message : "Tồn tại trang web khác có cùng url"
                 });
@@ -84,7 +84,7 @@ class SourceRouter   {
         }
         catch (err) {
             console.log(err);
-            res.status(400).json({
+            res.status(500).json({
                 success: false,
                 error : err,
             })
@@ -97,14 +97,23 @@ class SourceRouter   {
     async deleteSource (req, res, next) {
         const { sourceId } = req.params;
         try {
-            await Source.findByIdAndDelete( sourceId );
-            res.json({
-                success : true,
-                message : 'Delete successfully'
-            })
+            const articles = await Article.find({ source : sourceId });
+            if(articles.length !== 0) {
+                res.status(400).json({
+                    success: false,
+                    message : "Xóa hết bài báo thuộc nguồn này trước",
+                });
+            }
+            else {
+                await Source.findByIdAndDelete( sourceId );
+                res.json({
+                    success : true,
+                    message : 'Delete successfully'
+                })
+            }
         }
         catch (err) {
-            res.status(400).json({
+            res.status(500).json({
                 success: false,
                 error : err,
             });
